requirements.txt
# facebook_business_bot/requirements.txt
python-telegram-bot==20.8
requests==2.31.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0